+++ dfm-sideline-sidekick-app/components/barStyles.tsx
@@ -1,0 +1,18 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+    bottomBar: {
+      flexDirection: "row",
+      justifyContent: "space-around",
+      alignItems: "center",
+      backgroundColor: "#f0f0f0",
+      padding: 8,
+      position: "absolute",
+      bottom: 0,
+      left: 0,
+      right: 0,
+      height: 79,
+    },
+  });
+
+  export default styles;
