--- dfm-sideline-sidekick-app/pages/HomePage.tsx
@@ -1,11 +1,0 @@
-import { Text, View } from "react-native";
-
-export const HomePage = () => {
-  return (
-    <View style={{ flex: 1, alignItems: "center", justifyContent: "center" }}>
-      <Text>Home Dummy Page</Text>
-    </View>
-  );
-};
-
-export default HomePage;
